Refactor duplicated camera code in CameraManip
