+++ app/src/main/java/com/github/polybooks/TakeBookPictureActivity.kt
@@ -1,9 +1,7 @@
 package com.github.polybooks
 
-import android.Manifest
 import android.content.Context
 import android.content.Intent
-import android.content.pm.PackageManager
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
 import android.os.Bundle
@@ -11,10 +9,16 @@ import android.util.Log
 import android.view.View
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
-import androidx.camera.core.*
-import androidx.camera.lifecycle.ProcessCameraProvider
+import androidx.camera.core.ImageCapture
+import androidx.camera.core.ImageCaptureException
+import androidx.camera.core.ImageProxy
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
+import com.github.polybooks.utils.CameraManip.REQUEST_CODE_PERMISSIONS
+import com.github.polybooks.utils.CameraManip.REQUIRED_PERMISSIONS
+import com.github.polybooks.utils.CameraManip.TAG
+import com.github.polybooks.utils.CameraManip.allPermissionsGranted
+import com.github.polybooks.utils.CameraManip.startCamera
 import kotlinx.android.synthetic.main.activity_take_book_picture.*
 import java.io.ByteArrayOutputStream
 import java.io.FileOutputStream
@@ -38,16 +42,19 @@ class TakeBookPictureActivity : AppCompatActivity() {
         stringISBN = intent.getStringExtra(EXTRA_ISBN)
 
         // Request camera permissions
-        if (allPermissionsGranted()) {
-            startCamera()
+        if (allPermissionsGranted(baseContext)) {
+            startTakeBookPictureCamera()
         } else {
             ActivityCompat.requestPermissions(
                 this, REQUIRED_PERMISSIONS, REQUEST_CODE_PERMISSIONS)
         }
 
-        // Set up the listener for take photo button
-
         cameraExecutor = Executors.newSingleThreadExecutor()
+    }
+
+    private fun startTakeBookPictureCamera() {
+        imageCapture = ImageCapture.Builder().build()
+        startCamera(this, viewFinder, imageCapture!!)
     }
 
     fun takePhoto(view: View) {
@@ -108,55 +115,14 @@ class TakeBookPictureActivity : AppCompatActivity() {
         return pictureFileName
     }
 
-    private fun startCamera() {
-        val cameraProviderFuture = ProcessCameraProvider.getInstance(this)
-
-        cameraProviderFuture.addListener({
-            // Used to bind the lifecycle of cameras to the lifecycle owner
-            val cameraProvider: ProcessCameraProvider = cameraProviderFuture.get()
-
-            // Preview
-            val preview = Preview.Builder()
-                .build()
-                .also {
-                    it.setSurfaceProvider(viewFinder.surfaceProvider)
-                }
-
-            imageCapture = ImageCapture.Builder().build()
-
-            // Select back camera as a default
-            val cameraSelector = CameraSelector.DEFAULT_BACK_CAMERA
-            // TODO Other option:
-            //       = CameraSelector.Builder().requireLensFacing(CameraSelector.LENS_FACING_BACK).build()
-
-            try {
-                // Unbind use cases before rebinding
-                cameraProvider.unbindAll()
-
-                // Bind use cases to camera
-                cameraProvider.bindToLifecycle(
-                    this, cameraSelector, preview, imageCapture)
-
-            } catch(exc: Exception) {
-                Log.e(TAG, "Use case binding failed", exc)
-            }
-
-        }, ContextCompat.getMainExecutor(this))
-    }
-
-
-    private fun allPermissionsGranted() = REQUIRED_PERMISSIONS.all {
-        ContextCompat.checkSelfPermission(
-            baseContext, it) == PackageManager.PERMISSION_GRANTED
-    }
 
     override fun onRequestPermissionsResult(
         requestCode: Int, permissions: Array<String>, grantResults:
         IntArray) {
         super.onRequestPermissionsResult(requestCode, permissions, grantResults)
         if (requestCode == REQUEST_CODE_PERMISSIONS) {
-            if (allPermissionsGranted()) {
-                startCamera()
+            if (allPermissionsGranted(baseContext)) {
+                startTakeBookPictureCamera()
             } else {
                 Toast.makeText(this,
                     "You must grant camera permissions to take a picture.",
@@ -171,10 +137,5 @@ class TakeBookPictureActivity : AppCompatActivity() {
         cameraExecutor.shutdown()
     }
 
-    companion object {
-        private const val TAG = "CameraXBasic"
-        private const val REQUEST_CODE_PERMISSIONS = 10
-        private val REQUIRED_PERMISSIONS = arrayOf(Manifest.permission.CAMERA)
-    }
 
 }
